Experimental attempt to reduce duplicate LDAP loading
Add a LDAPObjectManager which intends to act as a cache of all the
loaded LDAP items we have. This is then woven through the construction
of all our objects in order to ensure that they query back to it
when trying to access other instances.

The primary intended use for this is in the construction of details
dictionaries, particularly the one for colleges. Unfortunately even
this proves to be less useful than anticipated due to the way that
we currently construct teams for a college (a college's teams are
the set of teams of which any members of the college are also members).

We could change that such that we gathered a list of all teams and
then compared the members of the college to the members of each
team, but there's no guaruntee that that would be any faster (and
it would mean that we'd essentially load the whole of our LDAP store
into Python to do so).

As a result this commit is unlikely to make it into master, but may
be useful as a record of why the approach isn't very useful.

Change-Id: I230e0a5f968cb4cdd008246aae5281b73d23577c